Modify per review of yrezgui@
