+++ lite/examples/sound_classification/android_compose/app/src/main/AndroidManifest.xml
@@ -10,12 +10,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.JetSoundClassifier">
+        android:theme="@style/Theme.SoundClassifier">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.JetSoundClassifier.NoActionBar">
+            android:theme="@style/Theme.SoundClassifier.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
