+++ lite/examples/sound_classification/android_compose/app/src/main/java/org/tensorflow/lite/examples/soundclassifier/compose/ui/theme/Theme.kt
@@ -20,7 +20,7 @@ private val LightColorPalette = lightColors(
 )
 
 @Composable
-fun JetSoundClassifierTheme(
+fun SoundClassifierTheme(
   darkTheme: Boolean = isSystemInDarkTheme(),
   content: @Composable() () -> Unit
 ) {
