+++ lite/examples/sound_classification/android_compose/app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">JetSoundClassifier</string>
+    <string name="app_name">SoundClassifier</string>
     <string name="label_input">Input:</string>
     <string name="label_classification_interval">Classification\nInterval</string>
     <string name="tensorflow_lite_logo">TensorFlow Lite Logo</string>
